Refactored ini fixtures
