+++ Fritz.ConfigurationBuilders/Fritz.ConfigurationBuilders.csproj
@@ -45,6 +45,7 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="IniConfigurationBuilder.cs" />
+    <Compile Include="KeyValueMode.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
   <ItemGroup>
