+++ Fritz.ConfigurationBuilders/KeyValueMode.cs
@@ -1,0 +1,9 @@
+﻿namespace Fritz.ConfigurationBuilders
+{
+	public enum KeyValueMode
+	{
+		Strict,
+		Greedy
+	}
+
+}
